requirements.txt
numpy==1.23.5
pandas==1.5.2
plotly==5.24.1
plotly-express==0.4.1
python-dotenv==1.0.1
scikit-learn==1.2.1
spotipy==2.24.0
streamlit==1.40.1
setuptools==67.6.1
